+++ src/WhyDidIReboot/SettingsWindow.xaml.cs
@@ -1,0 +1,15 @@
+using System.Windows;
+
+namespace WhyDidIReboot;
+
+public partial class SettingsWindow : Window
+{
+    public SettingsWindow(MainViewModel vm)
+    {
+        InitializeComponent();
+        DataContext = vm;
+        SourceInitialized += (_, _) => ThemeManager.ApplyTitleBar(this);
+    }
+
+    private void Close_Click(object sender, RoutedEventArgs e) => Close();
+}